added users stepone.vue
